# release-22.2: opt: speed up lookup constraint builder