--- a/pkg/sql/opt/constraint/constraint_set_test.go
+++ b/pkg/sql/opt/constraint/constraint_set_test.go
@@ -361,7 +361,45 @@ func TestExtractConstColsForSet(t *testing.T) {
 	}
 }
 
-func TestHasSingleColumnConstValues(t *testing.T) {
+func TestHasSingleColumnNonNullConstValues(t *testing.T) {
+	type testCase struct {
+		constraints []string
+		col         opt.ColumnID
+		expected    bool
+	}
+	cases := []testCase{
+		{[]string{`/1: [/10 - /10]`}, 1, true},
+		{[]string{`/-1: [/10 - /10]`}, 1, true},
+		{[]string{`/1: [/10 - /11]`}, 0, false},
+		{[]string{`/1: [/10 - /10] [/11 - /11]`}, 1, true},
+		{[]string{`/1: [/10 - /10] [/11 - /11] [/12 - /12]`}, 1, true},
+		{[]string{`/1: [/10 - /10] [/11 - /11] [/12 - /13]`}, 0, false},
+		{[]string{`/1: [/NULL - /NULL]`}, 0, false},
+		{[]string{`/1: [/NULL - /NULL] [/10 - /10] [/11 - /11]`}, 0, false},
+		{[]string{`/1/2: [/10/2 - /10/4]`}, 0, false},
+		{[]string{`/1/2: [/10/2 - /10/2]`}, 0, false},
+		{[]string{`/1: [/10 - /10]`, `/2: [/8 - /8]`}, 0, false},
+		{[]string{`/1: [/10 - /10]`, `/2: [/8 - /8]`}, 0, false},
+		{[]string{`/1: [/10 - /10]`, `/1/2: [/10/8 - /10/8]`}, 0, false},
+		{[]string{`/1: [/10 - /10]`, `/1/2: [/10/8 - /10/8]`}, 0, false},
+	}
+	st := cluster.MakeTestingClusterSettings()
+	evalCtx := eval.NewTestingEvalContext(st)
+	for _, tc := range cases {
+		cs := Unconstrained
+		for _, constraint := range tc.constraints {
+			constraint := ParseConstraint(evalCtx, constraint)
+			cs = cs.Intersect(evalCtx, SingleConstraint(&constraint))
+		}
+		col, res := cs.HasSingleColumnNonNullConstValues(evalCtx)
+		if res != tc.expected || col != tc.col {
+			t.Errorf("%s: expected %t,%d got %t,%d", cs, tc.expected, tc.col, res, col)
+
+		}
+	}
+}
+
+func TestExtractSingleColumnNonNullConstValues(t *testing.T) {
 	type testCase struct {
 		constraints []string
 		col         opt.ColumnID
@@ -374,6 +412,8 @@ func TestHasSingleColumnConstValues(t *testing.T) {
 		{[]string{`/1: [/10 - /10] [/11 - /11]`}, 1, []int{10, 11}},
 		{[]string{`/1: [/10 - /10] [/11 - /11] [/12 - /12]`}, 1, []int{10, 11, 12}},
 		{[]string{`/1: [/10 - /10] [/11 - /11] [/12 - /13]`}, 0, nil},
+		{[]string{`/1: [/NULL - /NULL]`}, 0, nil},
+		{[]string{`/1: [/NULL - /NULL] [/10 - /10] [/11 - /11]`}, 0, nil},
 		{[]string{`/1/2: [/10/2 - /10/4]`}, 0, nil},
 		{[]string{`/1/2: [/10/2 - /10/2]`}, 0, nil},
 		{
@@ -399,7 +439,7 @@ func TestHasSingleColumnConstValues(t *testing.T) {
 			constraint := ParseConstraint(evalCtx, constraint)
 			cs = cs.Intersect(evalCtx, SingleConstraint(&constraint))
 		}
-		col, vals, _ := cs.HasSingleColumnConstValues(evalCtx)
+		col, vals, _ := cs.ExtractSingleColumnNonNullConstValues(evalCtx)
 		var intVals []int
 		for _, val := range vals {
 			intVals = append(intVals, int(*val.(*tree.DInt)))